[feat] Molecules folder
