+++ src/components/charts/BarChart.tsx
@@ -3,14 +3,14 @@ import React, { useEffect, useRef, useState, useContext } from "react";
 import { Bar } from "react-chartjs-2";
 import { useNavigate } from "react-router";
 import useScreenSize from "../../hooks/useScreenSize";
-import CustomButton from "../Button";
-import FilteredCard from "../FilteredCard";
-import CustomCalendar from "../common/CustomCalendar";
-import CustomDropdown from "../DropDown";
+import CustomButton from "../atoms/Button";
+import FilteredCard from "../molecules/FilteredCard";
+import CustomCalendar from "../molecules/CustomCalendar";
+import CustomDropdown from "../atoms/DropDown";
 import CustomIcon from "../atoms/CustomIcon";
 import CustomImage from "../atoms/customimage";
 import CustomTab from "../atoms/customtab";
-import Loader from "../loader";
+import Loader from "../atoms/loader";
 import ChannelIcon from "../../assets/channel.svg";
 import FilterIcon from "../../assets/filter-dark.svg";
 import ArrowDownIcon from "../../assets/arrown_down_white.svg";
