+++ src/components/molecules/FilteredCard.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import CrossIcon from "../assets/cross.svg";
-import CustomButton from "./Button";
-import CustomImage from "./atoms/customimage";
-import { FilteredCardProps } from "../@types/components/commonTypes";
+import CustomButton from "../atoms/Button";
+import CustomImage from "../atoms/customimage";
+import { FilteredCardProps } from "../../@types/components/commonTypes";
 
 const FilteredCard: React.FC<FilteredCardProps> = (props) => {
   return (
