+++ src/components/molecules/orderstatuspopup.tsx
@@ -1,12 +1,12 @@
 import { Column } from "primereact/column";
 import { FC } from "react";
-import CustomDialog from "./atoms/customdialog";
-import CustomTable from "./atoms/customtable";
+import CustomDialog from "../atoms/customdialog";
+import CustomTable from "../atoms/customtable";
 import {
   ORDER_DETAILS_LABELS,
   ORDER_STATUS_LIST,
-} from "../constants/appConstants";
-import { getTableHeaders } from "../utils/Utils";
+} from "../../constants/appConstants";
+import { getTableHeaders } from "../../utils/Utils";
 
 type OrderStatusPopupProps = {
   openDialog: boolean;
