+++ src/components/organisms/promotionspopup.tsx
@@ -1,8 +1,8 @@
 import { FC } from "react";
-import CustomDialog from "./atoms/customdialog";
-import PromotionCard from "./common/promotioncard";
-import { PromotionsPopupProps } from "../@types/pages/OrderDetails";
-import { ORDER_DETAILS_LABELS } from "../constants/appConstants";
+import CustomDialog from "../atoms/customdialog";
+import PromotionCard from "../molecules/promotioncard";
+import { PromotionsPopupProps } from "../../@types/pages/OrderDetails";
+import { ORDER_DETAILS_LABELS } from "../../constants/appConstants";
 
 const PromotionsPopup: FC<PromotionsPopupProps> = (props) => {
   return (
