+++ src/components/organisms/sidepanegrid.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react";
 import { useNavigate } from "react-router-dom";
-import MenuCard from "./menucard";
+import MenuCard from "../molecules/menucard";
 import { SidePaneGridProps, IMenu } from "../../@types/components/commonTypes";
 
 const SidePaneGrid: FC<SidePaneGridProps> = (props) => {
@@ -14,8 +14,9 @@ const SidePaneGrid: FC<SidePaneGridProps> = (props) => {
 
   return (
     <div
-      className={`transition-all ease-in-out !duration-300 min-h-100vh-56 !w-[367px] ${props.showSidePaneGrid ? "left-0  right-full" : "-left-[367px] right-0"
-        } drop-shadow absolute z-10 bg-black-200`}
+      className={`transition-all ease-in-out !duration-300 min-h-100vh-56 !w-[367px] ${
+        props.showSidePaneGrid ? "left-0  right-full" : "-left-[367px] right-0"
+      } drop-shadow absolute z-10 bg-black-200`}
     >
       <div className="flex flex-wrap p-1">
         {props.menuList.map((menu) => (
