+++ src/stories/FilteredCard.stories.ts
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import "primereact/resources/primereact.min.css";
 import "primereact/resources/themes/lara-light-indigo/theme.css";
-import FilteredCard from "../components/FilteredCard";
+import FilteredCard from "../components/molecules/FilteredCard";
 import SmallCalendar from "../assets/calendar_small.svg";
 import "../index.css";
