+++ src/stories/LineChart.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from "@storybook/react";
 import "primereact/resources/primereact.min.css";
 import "primereact/resources/themes/lara-light-indigo/theme.css";
 import { BrowserRouter as Router } from "react-router-dom";
-import LineChart from "../components/LineChart";
+import LineChart from "../components/molecules/LineChart";
 
 const meta = {
   title: "Components/LineChart",
