+++ src/stories/menucard.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import MenuCard from "../components/common/menucard";
+import MenuCard from "../components/molecules/menucard";
 import DashboardIcon from "../assets/dashboard.svg";
 
 const meta = {
