+++ src/stories/navbar.stories.tsx
@@ -1,7 +1,7 @@
 import { INITIAL_VIEWPORTS } from "@storybook/addon-viewport";
 import type { Meta, StoryObj } from "@storybook/react";
 import { BrowserRouter as Router } from "react-router-dom";
-import Navbar from "../components/common/navbar";
+import Navbar from "../components/organisms/navbar";
 
 const meta = {
   title: "Stories/navbar",
