+++ src/views/OrderDetails.tsx
@@ -2,16 +2,16 @@ import { useState, useEffect } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import { Column } from "primereact/column";
 import useScreenSize from "../hooks/useScreenSize";
-import Loader from "../components/loader";
-import OrderStatus from "../components/orderstatus";
-import PromotionsPopup from "../components/promotionspopup";
-import OrderStatusPopup from "../components/orderstatuspopup";
+import Loader from "../components/atoms/loader";
+import OrderStatus from "../components/molecules/orderstatus";
+import PromotionsPopup from "../components/organisms/promotionspopup";
+import OrderStatusPopup from "../components/molecules/orderstatuspopup";
 import CustomImage from "../components/atoms/customimage";
 import CustomTable from "../components/atoms/customtable";
 import CustomIcon from "../components/atoms/CustomIcon";
 import Card from "../components/atoms/Card";
-import CustomButton from "../components/Button";
-import ItemInformationsPopUp from "../components/itemInformationsPopUp";
+import CustomButton from "../components/atoms/Button";
+import ItemInformationsPopUp from "../components/molecules/itemInformationsPopUp";
 import PromotionsIcon from "../../assets/promotions_white.svg";
 import OrderClockIcon from "../../assets/order_clock_white.svg";
 import RightArrowIcon from "../../assets/right_arrow.svg";
