+++ src/views/homePage.tsx
@@ -2,14 +2,14 @@ import { useContext, useEffect, useState } from "react";
 import useScreenSize from "../hooks/useScreenSize";
 import OPM from "./opm";
 import OpmComparison from "./opmComparison";
-import CustomButton from "../components/Button";
+import CustomButton from "../components/atoms/Button";
 import BarChart from "../components/charts/BarChart";
 import CustomImage from "../components/atoms/customimage";
-import HomeCard from "../components/common/homeCard";
-import GlobalLoader from "../components/globalLoader";
-import Loader from "../components/loader";
-import TimeTracker from "../components/timeTracker";
-import LoaderPortal from "../components/loaderPortal";
+import HomeCard from "../components/molecules/homeCard";
+import GlobalLoader from "../components/molecules/globalLoader";
+import Loader from "../components/atoms/loader";
+import TimeTracker from "../components/atoms/timeTracker";
+import LoaderPortal from "../components/atoms/loaderPortal";
 import avgOpmcompIcon from "../assets/avg_opm_comp.svg";
 import avgOrdersPerMinIcon from "../assets/avg_orders_per_min.svg";
 import infoIcon from "../assets/info_icon.svg";
